Delete now-unused data model code
