+++ data/model/oci/tag.py
@@ -21,6 +21,7 @@ from data.database import (
     Namespace,
     RepositoryNotification,
     ExternalNotificationEvent,
+    db_random_func,
 )
 from data.model.oci.shared import get_legacy_image_for_manifest
 from data.model import config
@@ -617,3 +618,81 @@ def lookup_notifiable_tags_for_legacy_image(docker_image_id, storage_uuid, event
         # If found in a repository with the valid event, yield the tag(s) that contains the image.
         for tag in tags_containing_legacy_image(image):
             yield tag
+
+
+def get_tags_for_legacy_image(image_id):
+    """ Returns the Tag's that have the associated legacy image. 
+    
+        NOTE: This is for legacy support in the old security notification worker and should
+        be removed once that code is no longer necessary.
+    """
+    return filter_to_alive_tags(
+        Tag.select()
+        .distinct()
+        .join(Manifest)
+        .join(ManifestLegacyImage)
+        .where(ManifestLegacyImage.image == image_id)
+    )
+
+
+def filter_has_repository_event(query, event):
+    """ Filters the query by ensuring the repositories returned have the given event.
+    
+        NOTE: This is for legacy support in the old security notification worker and should
+        be removed once that code is no longer necessary.
+    """
+    return (
+        query.join(Repository)
+        .join(RepositoryNotification)
+        .where(RepositoryNotification.event == event)
+    )
+
+
+def filter_tags_have_repository_event(query, event):
+    """ Filters the query by ensuring the tags live in a repository that has the given
+        event. Also orders the results by lifetime_start_ms.
+    
+        NOTE: This is for legacy support in the old security notification worker and should
+        be removed once that code is no longer necessary.
+    """
+    query = filter_has_repository_event(query, event)
+    query = query.switch(Tag).order_by(Tag.lifetime_start_ms.desc())
+    return query
+
+
+def find_repository_with_garbage(limit_to_gc_policy_s):
+    """ Returns a repository that has garbage (defined as an expired Tag that is past
+        the repo's namespace's expiration window) or None if none.
+    """
+    expiration_timestamp = get_epoch_timestamp_ms() - (limit_to_gc_policy_s * 1000)
+
+    try:
+        candidates = (
+            Tag.select(Tag.repository)
+            .join(Repository)
+            .join(Namespace, on=(Repository.namespace_user == Namespace.id))
+            .where(
+                ~(Tag.lifetime_end_ms >> None),
+                (Tag.lifetime_end_ms <= expiration_timestamp),
+                (Namespace.removed_tag_expiration_s == limit_to_gc_policy_s),
+            )
+            .limit(500)
+            .distinct()
+            .alias("candidates")
+        )
+
+        found = (
+            Tag.select(candidates.c.repository_id)
+            .from_(candidates)
+            .order_by(db_random_func())
+            .get()
+        )
+
+        if found is None:
+            return
+
+        return Repository.get(Repository.id == found.repository_id)
+    except Tag.DoesNotExist:
+        return None
+    except Repository.DoesNotExist:
+        return None
